move get ips and get hostname to workloadbase
