+++ src/workload.py
@@ -132,47 +132,3 @@ class EtcdWorkload(WorkloadBase):
     def enable_database(self) -> None:
         self.enable_service()
         self.start()
-
-    @override
-    def get_public_ip(self) -> str | None:
-        """Get the Public IP address of the current unit."""
-        cmd = "unit-get public-address"
-        try:
-            output = subprocess.run(
-                cmd,
-                check=True,
-                text=True,
-                shell=True,
-                capture_output=True,
-                timeout=10,
-            )
-        except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as e:
-            logger.error(f"Error executing command '{cmd}': {e}")
-            return None
-
-        if output.returncode != 0:
-            return None
-
-        return output.stdout.strip()
-
-    @override
-    def get_private_ip(self) -> str | None:
-        """Get the Private IP address of the current unit."""
-        cmd = "unit-get private-address"
-        try:
-            output = subprocess.run(
-                cmd,
-                check=True,
-                text=True,
-                shell=True,
-                capture_output=True,
-                timeout=10,
-            )
-        except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as e:
-            logger.error(f"Error executing command '{cmd}': {e}")
-            return None
-
-        if output.returncode != 0:
-            return None
-
-        return output.stdout.strip()
